Add files via upload
This year’s talks between Serbia and Kosovo organized in Berlin in January, Brussels in July and Washington in September proved a new engagement by the Western oriented international community in relation to the long-year frozen dialogue between Belgrade and Pristina. The Berlin talks, held at the US embassy in January this year, left an impression that the USA are giving “new life” to the normalization process of Serbia and Kosovo. The US were not ready to leave the dialogue entirely to the EU. However, this new economic meeting had the same flaws as the EU facilitated one – it lacked transparency and clarity. Despite that, the attempt to revive the Belgrade-Pristina dialogue giving it an economic impetus could turn out to be productive, as it leaves the heavy political questions for to be resolved at a later stage.